Rename to dffml-model-scikit
